+++ lib/archetype/sass_extensions/functions/styleguide/components.rb
@@ -12,14 +12,23 @@ module Archetype::SassExtensions::Styleguide
   # - {Boolean} whether or not the component was added
   #
   def styleguide_add_component(id, data, default = nil, theme = nil, force = false)
+    debug = Compass.configuration.styleguide_debug
+    helpers.debug "[archetype:styleguide] attempting to register component for `#{id}`" if debug
     _styleguide_mutex_helper(id, theme) do |id, theme|
       components = theme[:components]
       # if force was true, we have to invalidate the memoizer
       memoizer.clear(theme[:name]) if force
       # if we already have the component, don't create it again
-      return bool(false) if component_exists?(id, theme, nil, force) || component_is_frozen?(id, theme)
+      if component_exists?(id, theme, nil, force) || component_is_frozen?(id, theme)
+        helpers.debug "[archetype:styleguide] skipping component registration for `#{id}`. the component is already registered or frozen" if debug
+        return bool(false)
+      end
       # otherwise add it
       components[id] = helpers.data_to_hash(default, 1, SPECIAL, ADDITIVES).merge(helpers.data_to_hash(data, 1, SPECIAL, ADDITIVES))
+      if debug
+        helpers.debug "[archetype:styleguide] successfully registered component `#{id}`"
+        helpers.debug components[id].inspect
+      end
       return bool(true)
     end
   end
@@ -40,16 +49,29 @@ module Archetype::SassExtensions::Styleguide
   # - {Boolean} whether or not the component was extended
   #
   def styleguide_extend_component(id, data, theme = nil, extension = nil, force = false)
+    debug = Compass.configuration.styleguide_debug
+    helpers.debug "[archetype:styleguide] attempting to extend component for `#{id}`" if debug
+    helpers.debug "[archetype:styleguide] extension name is `#{extension.to_sass}`" if (debug && !extension.nil?)
     _styleguide_mutex_helper(id, theme) do |id, theme|
       components = theme[:components]
       # if force was set, we'll create a random token for the name
-      extension = random_uid('extension') if force
+      if force
+        extension = random_uid('extension')
+        helpers.debug "[archetype:styleguide] forcibly extending..." if debug
+      end
       # use the extension name or a snapshot of the extension
       extension = helpers.to_str(extension || data.to_sass)
       extensions = theme[:extensions]
-      return bool(false) if component_exists?(id, theme, extension, force) || component_is_frozen?(id, theme)
+      if component_exists?(id, theme, extension, force) || component_is_frozen?(id, theme)
+        helpers.debug "[archetype:styleguide] skipping component extension for `#{id}`. the extension is already registered or frozen" if debug
+        return bool(false)
+      end
       extensions.push(extension)
       components[id] = (components[id] ||= Archetype::Hash.new).rmerge(helpers.data_to_hash(data, 1, SPECIAL, ADDITIVES))
+        if debug
+          helpers.debug "[archetype:styleguide] successfully extended component `#{id}`"
+          helpers.debug components[id].inspect
+        end
       return bool(true)
     end
   end
@@ -89,6 +111,7 @@ module Archetype::SassExtensions::Styleguide
   # - <tt>$theme</tt> {String} the theme to insert the component into
   #
   def styleguide_remove_component(id, theme = nil)
+    helpers.debug "[archetype:styleguide] removing component `#{id}`" if Compass.configuration.styleguide_debug
     _styleguide_mutex_helper(id, theme) do |id, theme|
       theme[:components].delete(id)
       theme[:extensions].push(random_uid('remove'))
@@ -106,6 +129,7 @@ module Archetype::SassExtensions::Styleguide
   # - <tt>$theme</tt> {String} the theme to insert the component into
   #
   def styleguide_freeze_component(id, theme = nil)
+    helpers.debug "[archetype:styleguide] freezing component `#{id}`" if Compass.configuration.styleguide_debug
     _styleguide_mutex_helper(id, theme) do |id, theme|
       theme[:frozen] << id
       return bool(true)
@@ -121,6 +145,7 @@ module Archetype::SassExtensions::Styleguide
   # - <tt>$theme</tt> {String} the theme to insert the component into
   #
   def styleguide_freeze_all_components(theme = nil)
+    helpers.debug "[archetype:styleguide] freezing ALL components" if Compass.configuration.styleguide_debug
     _styleguide_mutex_helper do
       theme = get_theme(theme)
       theme[:frozen] = Set.new(theme[:components].keys)
@@ -137,6 +162,7 @@ module Archetype::SassExtensions::Styleguide
   # - <tt>$theme</tt> {String} the theme to insert the component into
   #
   def styleguide_thaw_component(id, theme = nil)
+    helpers.debug "[archetype:styleguide] thawing component `#{id}`" if Compass.configuration.styleguide_debug
     _styleguide_mutex_helper(id, theme) do |id, theme|
       theme[:frozen].delete(id)
       return bool(true)
@@ -152,6 +178,7 @@ module Archetype::SassExtensions::Styleguide
   # - <tt>$theme</tt> {String} the theme to insert the component into
   #
   def styleguide_thaw_all_components(theme = nil)
+    helpers.debug "[archetype:styleguide] thawing ALL components" if Compass.configuration.styleguide_debug
     _styleguide_mutex_helper do
       theme = get_theme(theme)
       theme[:frozen].clear